feat: Add error handling to MessageController#create
This commit adds error handling to the `MessageController#add_message` method. If there is an error while adding a message to the database, the appropriate HTTP status code and error message are returned in the response. Additionally, the newly created message ID is included in the response body when the message is successfully added.